802.11n and dual antenna IQ capture and other fixes

+++ boards/sdk_update.sh
@@ -20,7 +20,11 @@ fi
 
 cp $BOARD_NAME/openwifi_$BOARD_NAME/openwifi_$BOARD_NAME.sdk/system_top_hw_platform_0 $BOARD_NAME/sdk/ -rf
 git add $BOARD_NAME/sdk/*
-cp $BOARD_NAME/openwifi_$BOARD_NAME/openwifi_$BOARD_NAME.runs/impl_1/system_top.ltx $BOARD_NAME/sdk/
+if [ -f "$BOARD_NAME/openwifi_$BOARD_NAME/openwifi_$BOARD_NAME.runs/impl_1/system_top.ltx" ]; then
+    cp $BOARD_NAME/openwifi_$BOARD_NAME/openwifi_$BOARD_NAME.runs/impl_1/system_top.ltx $BOARD_NAME/sdk/
+else
+    echo "No debug probe file found."
+fi
 git add $BOARD_NAME/sdk/*
